+++ fair/cli.py
@@ -104,7 +104,7 @@ def create(debug, output: str) -> None:
 @click.option(
     "--registry",
     help="Specify registry directory",
-    default=fdp_com.DEFAULT_REGISTRY_LOCATION,
+    default=None,
     show_default=True,
 )
 @click.option(